Background change
in this i added different backgrounds. after warp speed your background
may change so yoor in a different galaxy

+++ OOP_Assignment__1/OOP_Assignment__1.pde
@@ -47,7 +47,7 @@ ArrayList<Vital> vitals = new ArrayList<Vital>();
 
 int change = 0;
 float warp_back;
-int background = 1;
+float background = 1;
 
 void draw()
 {
